refactor(billing): remove /payment page in favor of /billing
The /payment page was a duplicate of billing functionality, causing users
to miss features like auto credit reload and invoices. All "Add Funds"
links now navigate to /billing with an openPayment query param that
auto-opens the payment popup.
