+++ apps/deploy-web/src/components/shared/TrialDeploymentTooltip.tsx
@@ -68,7 +68,10 @@ const AddFunds = ({ dependencies: d = DEPENDENCIES }: { dependencies: typeof DEP
   return (
     <div className="flex flex-col gap-2">
       <p className="text-xs">Add funds to activate your account.</p>
-      <d.AddFundsLink className={cn("w-full space-x-2 hover:no-underline", buttonVariants({ variant: "default" }))} href={urlService.payment()}>
+      <d.AddFundsLink
+        className={cn("w-full space-x-2 hover:no-underline", buttonVariants({ variant: "default" }))}
+        href={urlService.billing({ openPayment: true })}
+      >
         <HandCard className="text-xs" />
         <span className="whitespace-nowrap">Add Funds</span>
       </d.AddFundsLink>
