--- apps/deploy-web/src/components/user/payment/index.ts
@@ -1,3 +1,0 @@
-export * from "./AddPaymentMethodPopup";
-export * from "./DeletePaymentMethodPopup";
-export * from "./PaymentForm";
