--- apps/deploy-web/src/pages/payment.tsx
@@ -1,352 +1,0 @@
-import React, { useEffect, useRef, useState } from "react";
-import { Alert, Button, Snackbar } from "@akashnetwork/ui/components";
-import { Xmark } from "iconoir-react";
-import { useTheme } from "next-themes";
-import { useSnackbar } from "notistack";
-
-import Layout from "@src/components/layout/Layout";
-import OnboardingRedirect from "@src/components/onboarding/OnboardingRedirect/OnboardingRedirect";
-import { ThreeDSecurePopup } from "@src/components/shared/PaymentMethodForm/ThreeDSecurePopup";
-import { PaymentMethodsList } from "@src/components/shared/PaymentMethodsList";
-import { Title } from "@src/components/shared/Title";
-import { AddPaymentMethodPopup, DeletePaymentMethodPopup, PaymentForm } from "@src/components/user/payment";
-import { PaymentSuccessAnimation } from "@src/components/user/payment/PaymentSuccessAnimation";
-import { usePaymentPolling } from "@src/context/PaymentPollingProvider";
-import { useSettings } from "@src/context/SettingsProvider";
-import { useWallet } from "@src/context/WalletProvider";
-import { Guard } from "@src/hoc/guard/guard.hoc";
-import { use3DSecure } from "@src/hooks/use3DSecure";
-import { useIsOnboarded } from "@src/hooks/useIsOnboarded";
-import { useNotificator } from "@src/hooks/useNotificator";
-import { useUser } from "@src/hooks/useUser";
-import { defineServerSideProps } from "@src/lib/nextjs/defineServerSideProps/defineServerSideProps";
-import { redirectIfAccessTokenExpired } from "@src/lib/nextjs/pageGuards/pageGuards";
-import { usePaymentMethodsQuery, usePaymentMutations, useSetupIntentMutation } from "@src/queries";
-import { handleCouponError, handleStripeError } from "@src/utils/stripeErrorHandler";
-
-const MINIMUM_PAYMENT_AMOUNT = 20;
-
-const PayPage: React.FunctionComponent = () => {
-  const { resolvedTheme } = useTheme();
-  const [amount, setAmount] = useState<string>("");
-  const [coupon, setCoupon] = useState<string>("");
-  const [selectedPaymentMethodId, setSelectedPaymentMethodId] = useState<string | undefined>();
-  const [showAddPaymentMethod, setShowAddPaymentMethod] = useState(false);
-  const [showDeleteConfirmation, setShowDeleteConfirmation] = useState(false);
-  const [cardToDelete, setCardToDelete] = useState<string>();
-  const [amountError, setAmountError] = useState<string>();
-  const [showPaymentSuccess, setShowPaymentSuccess] = useState<{ amount: string; show: boolean }>({ amount: "", show: false });
-  const [error, setError] = useState<string>();
-  const [errorAction, setErrorAction] = useState<string>();
-  const submittedAmountRef = useRef<string>("");
-  const isDarkMode = resolvedTheme === "dark";
-  const { enqueueSnackbar } = useSnackbar();
-  const notificator = useNotificator();
-  const { user } = useUser();
-  const { data: paymentMethods = [], isLoading: isLoadingPaymentMethods, refetch: refetchPaymentMethods } = usePaymentMethodsQuery();
-  const { data: setupIntent, mutate: createSetupIntent, reset: resetSetupIntent } = useSetupIntentMutation();
-  const {
-    confirmPayment: { isPending: isConfirmingPayment, mutateAsync: confirmPayment },
-    applyCoupon: { isPending: isApplyingCoupon, mutateAsync: applyCoupon },
-    removePaymentMethod
-  } = usePaymentMutations();
-  const { pollForPayment, isPolling } = usePaymentPolling();
-  const threeDSecure = use3DSecure({
-    onSuccess: () => {
-      pollForPayment();
-      setShowPaymentSuccess({ amount: submittedAmountRef.current, show: true });
-      setAmount("");
-      setCoupon("");
-    },
-    showSuccessMessage: false
-  });
-  const { settings } = useSettings();
-
-  const isLoading = isLoadingPaymentMethods;
-  const { isTrialing } = useWallet();
-
-  useEffect(() => {
-    if (paymentMethods.length > 0) {
-      if (!selectedPaymentMethodId || !paymentMethods.some(method => method.id === selectedPaymentMethodId)) {
-        setSelectedPaymentMethodId(paymentMethods[0].id);
-      }
-    }
-  }, [paymentMethods, selectedPaymentMethodId]);
-
-  useEffect(() => {
-    if (amount) {
-      validateAmount(parseFloat(amount));
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [amount]);
-
-  const clearError = () => {
-    if (error) {
-      setError(undefined);
-      setErrorAction(undefined);
-    }
-  };
-
-  const handlePayment = async (paymentMethodId: string) => {
-    if (!amount) return;
-    if (!selectedPaymentMethodId || !paymentMethods.some(method => method.id === selectedPaymentMethodId)) return;
-
-    // Capture the submitted amount before starting the payment flow
-    submittedAmountRef.current = amount;
-    clearError();
-
-    try {
-      const response = await confirmPayment({
-        userId: user?.id || "",
-        paymentMethodId,
-        amount: parseFloat(amount),
-        currency: "usd"
-      });
-
-      if (response && response.requiresAction && response.clientSecret && response.paymentIntentId) {
-        threeDSecure.start3DSecure({
-          clientSecret: response.clientSecret,
-          paymentIntentId: response.paymentIntentId,
-          paymentMethodId
-        });
-      } else if (response.success) {
-        pollForPayment();
-        setShowPaymentSuccess({ amount: submittedAmountRef.current, show: true });
-        setAmount("");
-        setCoupon("");
-      } else {
-        throw new Error("Payment failed");
-      }
-    } catch (error: unknown) {
-      console.error("Payment confirmation failed:", error);
-
-      const errorInfo = handleStripeError(error);
-
-      setError(errorInfo.message);
-      setErrorAction(errorInfo.userAction);
-      notificator.error(errorInfo.message, { title: "Payment Failed" });
-    }
-  };
-
-  const handleAddCardSuccess = async () => {
-    setShowAddPaymentMethod(false);
-    refetchPaymentMethods();
-  };
-
-  const handleShowAddPaymentMethod = () => {
-    resetSetupIntent();
-    createSetupIntent();
-    setShowAddPaymentMethod(true);
-  };
-
-  const handleClaimCoupon = async () => {
-    if (!coupon) return;
-
-    try {
-      const response = await applyCoupon({ coupon, userId: user?.id || "" });
-
-      if (response.error) {
-        const errorInfo = handleCouponError(response);
-        notificator.error(errorInfo.message, { title: "Payment Failed" });
-        return;
-      }
-
-      if (response.amountAdded && response.amountAdded > 0) {
-        pollForPayment();
-        setShowPaymentSuccess({ amount: response.amountAdded.toString(), show: true });
-      }
-
-      enqueueSnackbar(<Snackbar title="Coupon applied successfully!" iconVariant="success" />, { variant: "success", autoHideDuration: 5_000 });
-      setCoupon("");
-    } catch (error: unknown) {
-      const errorInfo = handleStripeError(error);
-      notificator.error(errorInfo.message, { title: "Payment Failed" });
-      console.error("Coupon application error:", error);
-    }
-  };
-
-  const handleRemovePaymentMethod = async (paymentMethodId: string) => {
-    setCardToDelete(paymentMethodId);
-    setShowDeleteConfirmation(true);
-  };
-
-  const confirmRemovePaymentMethod = async () => {
-    if (!cardToDelete) return;
-
-    try {
-      await removePaymentMethod.mutateAsync(cardToDelete);
-      setSelectedPaymentMethodId(undefined);
-      enqueueSnackbar(<Snackbar title="Payment method removed successfully" iconVariant="success" />, { variant: "success" });
-    } catch (error: unknown) {
-      console.error("Failed to remove payment method:", error);
-
-      const errorInfo = handleStripeError(error);
-      notificator.error(errorInfo.message, { title: "Payment Failed" });
-    } finally {
-      setShowDeleteConfirmation(false);
-      setCardToDelete(undefined);
-    }
-  };
-
-  const validateAmount = (value: number) => {
-    if (value <= 0) {
-      setAmountError("Amount must be greater than $0");
-      return false;
-    }
-
-    if (value < MINIMUM_PAYMENT_AMOUNT) {
-      setAmountError(`Minimum amount is $${MINIMUM_PAYMENT_AMOUNT}`);
-      return false;
-    }
-
-    setAmountError(undefined);
-    return true;
-  };
-
-  const handleAmountChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value;
-    setAmount(value);
-    clearError();
-
-    if (value !== "") {
-      validateAmount(parseFloat(value));
-    } else {
-      setAmountError(undefined);
-    }
-  };
-
-  const handleCouponChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value;
-    setCoupon(value);
-    clearError();
-  };
-
-  if (isLoading) {
-    return (
-      <Layout>
-        <div className="flex min-h-[50vh] items-center justify-center">
-          <div className="flex flex-col items-center gap-4">
-            <div className="h-8 w-8 animate-spin rounded-full border-4 border-primary border-t-transparent" />
-            <p className="text-muted-foreground">Loading payment information...</p>
-          </div>
-        </div>
-      </Layout>
-    );
-  }
-
-  return (
-    <Layout isLoading={isLoading}>
-      <div className="py-12">
-        <Title className="text-center">Payment Methods</Title>
-        <p className="mt-4 text-center text-gray-600">Manage your payment methods and make payments.</p>
-        <div className="mx-auto max-w-md py-6">
-          <PaymentSuccessAnimation
-            show={showPaymentSuccess.show}
-            amount={showPaymentSuccess.amount}
-            onComplete={() => setShowPaymentSuccess({ amount: "", show: false })}
-          />
-          <div className="mb-6">
-            <PaymentMethodsList
-              paymentMethods={paymentMethods}
-              isRemoving={removePaymentMethod.isPending}
-              onRemovePaymentMethod={handleRemovePaymentMethod}
-              isSelectable={true}
-              selectedPaymentMethodId={selectedPaymentMethodId}
-              onPaymentMethodSelect={setSelectedPaymentMethodId}
-              isTrialing={isTrialing}
-            />
-            <Button onClick={handleShowAddPaymentMethod} className="mt-6 w-full">
-              Add New Payment Method
-            </Button>
-          </div>
-
-          {paymentMethods.length > 0 && (
-            <div className="mt-6">
-              {settings.isBlockchainDown && (
-                <Alert variant="warning" className="mb-4">
-                  <p className="font-medium">
-                    We are currently experiencing a temporary blockchain outage, which may cause delays in processing your payments. Once the blockchain is back
-                    online, all pending transactions will be processed automatically.
-                    <br />
-                    If you encounter any issues or have urgent concerns, please don’t hesitate to reach out to us — we’re here to help.
-                  </p>
-                </Alert>
-              )}
-              <PaymentForm
-                amount={amount}
-                onAmountChange={handleAmountChange}
-                amountError={amountError}
-                coupon={coupon}
-                onCouponChange={handleCouponChange}
-                onClaimCoupon={handleClaimCoupon}
-                processing={isConfirmingPayment || isPolling}
-                selectedPaymentMethodId={selectedPaymentMethodId}
-                onPayment={handlePayment}
-                isApplyingCoupon={isApplyingCoupon}
-              />
-
-              {/* Show error inline if there's a critical error */}
-              {error && (
-                <div className="mx-auto mt-6 max-w-md">
-                  <Alert variant="destructive" className="mb-4">
-                    <p className="font-medium">Error Loading Payment Information</p>
-                    <p className="text-sm">{error}</p>
-                    {errorAction && (
-                      <p className="mt-2 text-sm text-muted-foreground">
-                        <strong>Suggestion:</strong> {errorAction}
-                      </p>
-                    )}
-                    <Button onClick={clearError} variant="default" size="sm" className="mt-2">
-                      <Xmark className="mr-2 h-4 w-4" />
-                      Clear Error
-                    </Button>
-                  </Alert>
-                </div>
-              )}
-            </div>
-          )}
-        </div>
-      </div>
-
-      <DeletePaymentMethodPopup
-        open={showDeleteConfirmation}
-        onClose={() => {
-          setShowDeleteConfirmation(false);
-          setCardToDelete(undefined);
-        }}
-        onConfirm={confirmRemovePaymentMethod}
-        isRemovingPaymentMethod={removePaymentMethod.isPending}
-      />
-
-      <AddPaymentMethodPopup
-        open={showAddPaymentMethod}
-        onClose={() => setShowAddPaymentMethod(false)}
-        clientSecret={setupIntent?.clientSecret}
-        isDarkMode={isDarkMode}
-        onSuccess={handleAddCardSuccess}
-      />
-
-      {threeDSecure.threeDSData?.clientSecret && (
-        <ThreeDSecurePopup
-          isOpen={threeDSecure.isOpen}
-          onSuccess={threeDSecure.handle3DSSuccess}
-          onError={threeDSecure.handle3DSError}
-          clientSecret={threeDSecure.threeDSData.clientSecret}
-          paymentIntentId={threeDSecure.threeDSData.paymentIntentId}
-          paymentMethodId={threeDSecure.threeDSData.paymentMethodId}
-          title="Payment Authentication"
-          description="Your bank requires additional verification for this payment."
-          successMessage="Payment authenticated successfully!"
-          errorMessage="Please try again or use a different payment method."
-        />
-      )}
-    </Layout>
-  );
-};
-
-export default Guard(PayPage, useIsOnboarded, OnboardingRedirect);
-
-export const getServerSideProps = defineServerSideProps({
-  if: redirectIfAccessTokenExpired,
-  route: "/payment"
-});
